Add a guide to deploying on Cloud Foundry
The protoype kit deploys on Cloud Foumdry without any modification and in user
research interviews for PaaS for Government people who have been using it for
prototyping mentioned that they'd had to search around a bit to work out how to
set the environment variables required by the kit.

This is an initial attempt to make something like the Heroku deployment guide
for Cloud Foundry so that the instructions are all there.

It was created by copying the Heroku guide and replacing the Heroku specific
language or instructions with the nearest equivalent for Cloud Foundry.

Because, unlike Heroku, there is more than one Cloud Foundry installation, some
extra concepts had to be introduced to explain that there are multiple Cloud
Foundry providers and that each one has its own domain for hosting apps on.
We've also had to make the signup steps more vague, as these will vary from
provider to provider.

The alternative to this would have been to only mention a single Cloud Foundry
provider, such as PaaS for Government. This would have made instructions
simpler, but made it slightly harder for people who used alternatives. Also,
PaaS for Government is currently in private beta and not ready for such
widespread use.